# docs/swagger.yaml
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  model.Order:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      orderDetail:
        items:
          $ref: '#/definitions/model.OrderDetail'
        type: array
      totalAmount:
        type: integer
      totalPrice:
        type: number
      updatedAt:
        type: string
      walletID:
        type: integer
    type: object
  model.OrderDetail:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      orderID:
        type: integer
      price:
        type: number
      productID:
        type: integer
      updatedAt:
        type: string
    type: object
...
  request.OrderItem:
    properties:
      amount:
        example: 10
        type: integer
      product_id:
        example: 1
        type: integer
    required:
    - amount
    - product_id
    type: object
  request.OrderRequest:
    properties:
      orders:
        items:
          $ref: '#/definitions/request.OrderItem'
        type: array
      wallet_id:
        example: 10
        type: integer
    required:
    - orders
    - wallet_id
    type: object
  request.PermissionInfoRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  request.ProductCategoryRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
...
  request.UpdatePermissionInfo:
    properties:
      description:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  request.UpdateProductCategory:
    properties:
      description:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  response.CreateDataResponse:
    properties:
      message:
        example: create success
        type: string
      response_key:
        type: string
      response_message:
        type: string
    type: object
  response.DeleteDataResponse:
    properties:
      message:
        example: delete success
        type: string
      response_key:
        type: string
      response_message:
        type: string
    type: object
  response.Order:
    properties:
      id:
        type: integer
      total_amount:
        type: integer
      total_price:
        type: number
    required:
    - total_amount
    - total_price
    type: object
  response.OrderPagination:
    properties:
      data:
        items:
          $ref: '#/definitions/response.Order'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      response_key:
        type: string
      response_message:
        type: string
      totalPage:
        type: integer
    type: object
  response.PermissionInfo:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  response.PermissionInfoPagination:
    properties:
      data:
        items:
          $ref: '#/definitions/response.PermissionInfo'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      response_key:
        type: string
      response_message:
        type: string
      totalPage:
        type: integer
    type: object
...
  response.ProductCategory:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  response.ProductCategoryPagination:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ProductCategory'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      response_key:
        type: string
      response_message:
        type: string
      totalPage:
        type: integer
    type: object
...
  response.UpdateDataResponse:
    properties:
      message:
        example: update success
        type: string
      response_key:
        type: string
      response_message:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /orders:
    get:
      description: Get Order List
      parameters:
      - description: int valid
        in: query
        name: page
        type: integer
      - description: int valid
        in: query
        name: pageSize
        type: integer
      - description: string valid
        in: query
        name: sortField
        type: string
      - description: string valid
        in: query
        name: sortValue
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OrderPagination'
      security:
      - Bearer: []
      summary: Get Order List
      tags:
      - Order
    post:
      description: Create Order
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.OrderRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Order'
      security:
      - Bearer: []
      summary: Create Order
      tags:
      - Order
  /permission_infos:
    get:
      description: Get PermissionInfo List
      parameters:
      - description: int valid
        in: query
        name: page
        type: integer
      - description: int valid
        in: query
        name: pageSize
        type: integer
      - description: string valid
        in: query
        name: sortField
        type: string
      - description: string valid
        in: query
        name: sortValue
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PermissionInfoPagination'
      security:
      - Bearer: []
      summary: Get PermissionInfo List
      tags:
      - PermissionInfo
    post:
      description: Create PermissionInfo
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.PermissionInfoRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CreateDataResponse'
      security:
      - Bearer: []
      summary: Create PermissionInfo
      tags:
      - PermissionInfo
  /permission_infos/{permissionInfoID}:
    delete:
      description: Delete PermissionInfo By Id
      parameters:
      - description: PermissionInfo ID
        in: path
        name: permissionInfoID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DeleteDataResponse'
      security:
      - Bearer: []
      summary: Delete PermissionInfo By Id
      tags:
      - PermissionInfo
    get:
      description: Get PermissionInfo By Id
      parameters:
      - description: PermissionInfo ID
        in: path
        name: permissionInfoID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PermissionInfo'
      security:
      - Bearer: []
      summary: Get PermissionInfo By Id
      tags:
      - PermissionInfo
    put:
      description: Update PermissionInfo By Id
      parameters:
      - description: PermissionInfo ID
        in: path
        name: permissionInfoID
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePermissionInfo'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UpdateDataResponse'
      security:
      - Bearer: []
      summary: Update PermissionInfo By Id
      tags:
      - PermissionInfo
...
  /products/categories:
    get:
      description: Get ProductCategory List
      parameters:
      - description: int valid
        in: query
        name: page
        type: integer
      - description: int valid
        in: query
        name: pageSize
        type: integer
      - description: string valid
        in: query
        name: sortField
        type: string
      - description: string valid
        in: query
        name: sortValue
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductCategoryPagination'
      security:
      - Bearer: []
      summary: Get ProductCategory List
      tags:
      - ProductCategory
    post:
      description: Create ProductCategory
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ProductCategoryRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CreateDataResponse'
      security:
      - Bearer: []
      summary: Create ProductCategory
      tags:
      - ProductCategory
  /products/categories/{productCategoryID}:
    delete:
      description: Delete ProductCategory By Id
      parameters:
      - description: User ID
        in: path
        name: productCategoryID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DeleteDataResponse'
      security:
      - Bearer: []
      summary: Delete ProductCategory By Id
      tags:
      - ProductCategory
    get:
      description: Get ProductCategory By Id
      parameters:
      - description: User ID
        in: path
        name: productCategoryID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ProductCategory'
      security:
      - Bearer: []
      summary: Get ProductCategory By Id
      tags:
      - ProductCategory
    put:
      description: Update ProductCategory By Id
      parameters:
      - description: User ID
        in: path
        name: productCategoryID
        required: true
        type: integer
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductCategory'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.UpdateDataResponse'
      security:
      - Bearer: []
      summary: Update ProductCategory By Id
      tags:
      - ProductCategory
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"